baseRoute bug fix

+++ src/CRUD.js
@@ -17,13 +17,29 @@ class CRUD {
         this.headers = headers;
     }
 
+    checkUrl(url) {
+        if (typeof url !== 'string') {
+            console.error('url not set')
+            return false;
+        }
+        return true;
+    }
+
     create(url, data) {
+
+        if (!this.checkUrl(url)) {
+            return new Promise();
+        }
         return axios.post(url, data, {
             headers: this.headers
         })
     }
 
     fetch(url, data) {
+
+        if (!this.checkUrl(url)) {
+            return new Promise();
+        }
         return axios.get(url, {
             headers: this.headers,
             params: data
@@ -31,12 +47,20 @@ class CRUD {
     }
 
     update(url, data) {
+
+        if (!this.checkUrl(url)) {
+            return new Promise();
+        }
         return axios.put(url, data, {
             headers: this.headers
         })
     }
 
     delete(url) {
+
+        if (!this.checkUrl(url)) {
+            return new Promise();
+        }
         return axios.delete(url, {
             headers: this.headers
         });
